UI for base device

+++ DomainData/Network.cs
@@ -3,7 +3,7 @@
     public class Network
     {
         public string IpAddress { get; set; }
-        public int SubnetMask { get; set; }
+        public byte SubnetMask { get; set; }
         public string Gateway { get; set; }
         public string DnsServer { get; set; }
         public string AlternativeDnsServer { get; set; }
